chore: abstract tilemap utils
