+++ packages/game/src/typedefs/TileData.ts
@@ -1,0 +1,13 @@
+// Module imports
+import {
+	Cell,
+	GridTile,
+	ImageTile,
+	Tilemap,
+} from 'pixi-tiled-loader'
+
+
+
+
+
+export type TileData = [Cell, GridTile | ImageTile, Tilemap]
